LPS-96040 I want to be able to encapsulate things needed by only 1 test, into the test
